Abstract dataset refactored into modules and package
