+++ autoarray/dataset/imaging/settings.py
@@ -1,7 +1,7 @@
 import logging
 from typing import List, Optional
 
-from autoarray.dataset.abstract_dataset import AbstractSettingsDataset
+from autoarray.dataset.abstract.settings import AbstractSettingsDataset
 from autoarray.structures.grids.uniform_2d import Grid2D
 
 logger = logging.getLogger(__name__)
