+++ autoarray/dataset/imaging/w_tilde.py
@@ -5,9 +5,9 @@ from typing import List, Optional
 
 from autoconf import cached_property
 
-from autoarray.dataset.abstract_dataset import AbstractWTilde
-from autoarray.dataset.abstract_dataset import AbstractSettingsDataset
-from autoarray.dataset.abstract_dataset import AbstractDataset
+from autoarray.dataset.abstract.w_tilde import AbstractWTilde
+from autoarray.dataset.abstract.settings import AbstractSettingsDataset
+from autoarray.dataset.abstract.dataset import AbstractDataset
 from autoarray.structures.arrays.uniform_2d import Array2D
 from autoarray.operators.convolver import Convolver
 from autoarray.structures.grids.uniform_2d import Grid2D
