+++ test_autoarray/dataset/abstract/test_dataset.py
@@ -1,8 +1,9 @@
 import logging
+import numpy as np
 
 import autoarray as aa
-from autoarray.dataset import abstract_dataset
-import numpy as np
+
+from autoarray.dataset.abstract import dataset as ds
 
 logger = logging.getLogger(__name__)
 
@@ -12,7 +13,7 @@ def test__dataset_takes_structures_of_different_formats():
     array = aa.Array1D.manual_native([1.0, 2.0], pixel_scales=1.0)
     noise_map = aa.Array1D.manual_native([1.0, 3.0], pixel_scales=1.0)
 
-    dataset = abstract_dataset.AbstractDataset(data=array, noise_map=noise_map)
+    dataset = ds.AbstractDataset(data=array, noise_map=noise_map)
 
     assert (dataset.data.native == np.array([1.0, 2.0])).all()
     assert (dataset.noise_map.native == np.array([1.0, 3.0])).all()
@@ -20,7 +21,7 @@ def test__dataset_takes_structures_of_different_formats():
     array = aa.Array2D.manual_native([[1.0, 2.0], [3.0, 4.0]], pixel_scales=1.0)
     noise_map = aa.Array2D.manual_native([[1.0, 2.0], [3.0, 5.0]], pixel_scales=1.0)
 
-    dataset = abstract_dataset.AbstractDataset(data=array, noise_map=noise_map)
+    dataset = ds.AbstractDataset(data=array, noise_map=noise_map)
 
     assert (dataset.data.native == np.array([[1.0, 2.0], [3.0, 4.0]])).all()
     assert (dataset.noise_map.native == np.array([[1.0, 2.0], [3.0, 5.0]])).all()
@@ -31,7 +32,7 @@ def test__inverse_noise_map():
     array = aa.Array2D.manual_native([[1.0, 2.0], [3.0, 4.0]], pixel_scales=1.0)
     noise_map = aa.Array2D.manual_native([[1.0, 2.0], [4.0, 8.0]], pixel_scales=1.0)
 
-    dataset = abstract_dataset.AbstractDataset(data=array, noise_map=noise_map)
+    dataset = ds.AbstractDataset(data=array, noise_map=noise_map)
 
     assert (
         dataset.inverse_noise_map.native == np.array([[1.0, 0.5], [0.25, 0.125]])
@@ -42,7 +43,7 @@ def test__signal_to_noise_map():
     array = aa.Array2D.manual_native([[1.0, 2.0], [3.0, 4.0]], pixel_scales=1.0)
     noise_map = aa.Array2D.manual_native([[10.0, 10.0], [30.0, 4.0]], pixel_scales=1.0)
 
-    dataset = abstract_dataset.AbstractDataset(data=array, noise_map=noise_map)
+    dataset = ds.AbstractDataset(data=array, noise_map=noise_map)
 
     assert (
         dataset.signal_to_noise_map.native == np.array([[0.1, 0.2], [0.1, 1.0]])
@@ -53,7 +54,7 @@ def test__signal_to_noise_map():
 
     noise_map = aa.Array2D.manual_native([[10.0, 10.0], [30.0, 4.0]], pixel_scales=1.0)
 
-    dataset = abstract_dataset.AbstractDataset(data=array, noise_map=noise_map)
+    dataset = ds.AbstractDataset(data=array, noise_map=noise_map)
 
     assert (
         dataset.signal_to_noise_map.native == np.array([[0.0, 0.2], [0.1, 0.0]])
@@ -67,7 +68,7 @@ def test__absolute_signal_to_noise_map():
 
     noise_map = aa.Array2D.manual_native([[10.0, 10.0], [30.0, 4.0]], pixel_scales=1.0)
 
-    dataset = abstract_dataset.AbstractDataset(data=array, noise_map=noise_map)
+    dataset = ds.AbstractDataset(data=array, noise_map=noise_map)
 
     assert (
         dataset.absolute_signal_to_noise_map.native
@@ -81,7 +82,7 @@ def test__potential_chi_squared_map():
     array = aa.Array2D.manual_native([[-1.0, 2.0], [3.0, -4.0]], pixel_scales=1.0)
     noise_map = aa.Array2D.manual_native([[10.0, 10.0], [30.0, 4.0]], pixel_scales=1.0)
 
-    dataset = abstract_dataset.AbstractDataset(data=array, noise_map=noise_map)
+    dataset = ds.AbstractDataset(data=array, noise_map=noise_map)
 
     assert (
         dataset.potential_chi_squared_map.native
@@ -99,10 +100,10 @@ def test__grid__uses_mask_and_settings(
     grid_2d_iterate_7x7,
 ):
 
-    dataset_1d = abstract_dataset.AbstractDataset(
+    dataset_1d = ds.AbstractDataset(
         data=aa.Array1D.manual_native(array=[1.0], pixel_scales=1.0),
         noise_map=aa.Array1D.manual_native(array=[1.0], pixel_scales=1.0),
-        settings=abstract_dataset.AbstractSettingsDataset(),
+        settings=ds.AbstractSettingsDataset(),
     )
 
     assert isinstance(dataset_1d.grid, aa.Grid1D)
@@ -115,20 +116,20 @@ def test__grid__uses_mask_and_settings(
         array=noise_map_7x7.native, mask=sub_mask_2d_7x7.mask_sub_1
     )
 
-    masked_imaging_7x7 = abstract_dataset.AbstractDataset(
+    masked_imaging_7x7 = ds.AbstractDataset(
         data=masked_image_7x7,
         noise_map=masked_noise_map_7x7,
-        settings=abstract_dataset.AbstractSettingsDataset(),
+        settings=ds.AbstractSettingsDataset(),
     )
 
     assert isinstance(masked_imaging_7x7.grid, aa.Grid2D)
     assert (masked_imaging_7x7.grid.binned == grid_2d_7x7).all()
     assert (masked_imaging_7x7.grid.slim == sub_grid_2d_7x7).all()
 
-    masked_imaging_7x7 = abstract_dataset.AbstractDataset(
+    masked_imaging_7x7 = ds.AbstractDataset(
         data=masked_image_7x7,
         noise_map=masked_noise_map_7x7,
-        settings=abstract_dataset.AbstractSettingsDataset(grid_class=aa.Grid2DIterate),
+        settings=ds.AbstractSettingsDataset(grid_class=aa.Grid2DIterate),
     )
 
     assert isinstance(masked_imaging_7x7.grid, aa.Grid2DIterate)
@@ -139,10 +140,10 @@ def test__grid_pixelization__uses_mask_and_settings(
     image_7x7, noise_map_7x7, sub_mask_2d_7x7, grid_2d_7x7, sub_grid_2d_7x7
 ):
 
-    masked_dataset_1d = abstract_dataset.AbstractDataset(
+    masked_dataset_1d = ds.AbstractDataset(
         data=aa.Array1D.manual_native(array=[1.0], pixel_scales=1.0),
         noise_map=aa.Array1D.manual_native(array=[1.0], pixel_scales=1.0),
-        settings=abstract_dataset.AbstractSettingsDataset(),
+        settings=ds.AbstractSettingsDataset(),
     )
 
     assert isinstance(masked_dataset_1d.grid, aa.Grid1D)
@@ -155,10 +156,10 @@ def test__grid_pixelization__uses_mask_and_settings(
         array=noise_map_7x7.native, mask=sub_mask_2d_7x7.mask_sub_1
     )
 
-    masked_imaging_7x7 = abstract_dataset.AbstractDataset(
+    masked_imaging_7x7 = ds.AbstractDataset(
         data=masked_image_7x7,
         noise_map=masked_noise_map_7x7,
-        settings=abstract_dataset.AbstractSettingsDataset(
+        settings=ds.AbstractSettingsDataset(
             grid_pixelization_class=aa.Grid2D, sub_size_pixelization=2
         ),
     )
@@ -167,10 +168,10 @@ def test__grid_pixelization__uses_mask_and_settings(
     assert (masked_imaging_7x7.grid_pixelization.binned == grid_2d_7x7).all()
     assert (masked_imaging_7x7.grid_pixelization.slim == sub_grid_2d_7x7).all()
 
-    masked_imaging_7x7 = abstract_dataset.AbstractDataset(
+    masked_imaging_7x7 = ds.AbstractDataset(
         data=masked_image_7x7,
         noise_map=masked_noise_map_7x7,
-        settings=abstract_dataset.AbstractSettingsDataset(
+        settings=ds.AbstractSettingsDataset(
             grid_pixelization_class=aa.Grid2D, sub_size=2, sub_size_pixelization=4
         ),
     )
@@ -181,10 +182,10 @@ def test__grid_pixelization__uses_mask_and_settings(
 
 def test__grid_settings__sub_size(image_7x7, noise_map_7x7):
 
-    dataset_7x7 = abstract_dataset.AbstractDataset(
+    dataset_7x7 = ds.AbstractDataset(
         data=image_7x7,
         noise_map=noise_map_7x7,
-        settings=abstract_dataset.AbstractSettingsDataset(
+        settings=ds.AbstractSettingsDataset(
             sub_size=1, sub_size_pixelization=1
         ),
     )
@@ -192,10 +193,10 @@ def test__grid_settings__sub_size(image_7x7, noise_map_7x7):
     assert dataset_7x7.grid.mask.sub_size == 1
     assert dataset_7x7.grid_pixelization.mask.sub_size == 1
 
-    dataset_7x7 = abstract_dataset.AbstractDataset(
+    dataset_7x7 = ds.AbstractDataset(
         data=image_7x7,
         noise_map=noise_map_7x7,
-        settings=abstract_dataset.AbstractSettingsDataset(
+        settings=ds.AbstractSettingsDataset(
             sub_size=2, sub_size_pixelization=2
         ),
     )
@@ -213,7 +214,7 @@ def test__new_imaging_with_arrays_trimmed_via_kernel_shape():
     )
     noise_map_array[4] = 2.0
 
-    dataset = abstract_dataset.AbstractDataset(data=data, noise_map=noise_map_array)
+    dataset = ds.AbstractDataset(data=data, noise_map=noise_map_array)
 
     dataset_trimmed = dataset.trimmed_after_convolution_from(kernel_shape=(3, 3))
